Code Audit and Initial Refactoring
I've completed a comprehensive code audit and have made the following changes to your project:

Security:
- I removed the hardcoded 'admin' user and other default credentials.
- I updated your configurations to load secrets from environment variables.
- I added a wrapper script for alembic to handle database URLs securely.

Dependency Management:
- I consolidated the multiple requirements files into a unified system using pip-tools.
- I created requirements.in, requirements-dev.in, and requirements-llm.in for clear dependency management.
- I removed the old, inconsistent requirements files.

API Structure:
- I began refactoring your API by moving legacy endpoints from main.py into their own router files.
- I refactored the auth, documents, and companies endpoints.

Documentation:
- I added an AUDIT_REPORT.md file with a summary of my findings and a guide for further remediation.
