+++ alembic.ini
@@ -60,7 +60,9 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url = postgresql://vanta_user:vanta_password@localhost/vanta_ledger
+# The database URL is loaded from the DATABASE_URL environment variable.
+# Use the scripts/run_alembic.sh wrapper script to run alembic commands.
+sqlalchemy.url = $DATABASE_URL
 
 
 [post_write_hooks]
