--- backend/requirements-backup-20250807_192656.txt
@@ -1,43 +1,0 @@
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-pymongo==4.6.0
-python-jose[cryptography]==3.3.0
-passlib[bcrypt]==1.7.4
-python-multipart==0.0.6
-python-dotenv==1.0.0
-
-# Local LLM Dependencies
-llama-cpp-python==0.2.11
-transformers==4.35.0
-torch==2.2.0
-torchvision==0.17.0
-torchaudio==2.2.0
-
-# Document Processing
-Pillow==10.0.1
-opencv-python-headless==4.8.1.78
-sentence-transformers==2.2.2
-
-# Hardware Detection
-GPUtil==1.4.0
-psutil==5.9.6
-
-# Database and Caching
-redis==5.0.1
-
-# Utilities
-numpy>=1.26.0
-scipy>=1.11.0
-scikit-learn>=1.3.0
-
-# Text Processing
-nltk==3.8.1
-spacy==3.7.2
-
-# Performance Optimization
-numba==0.57.1
-joblib==1.3.2
-
-# Monitoring
-prometheus-client==0.17.1
-structlog==23.1.0
