+++ backend/requirements-dev.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.12
 # by the following command:
 #
-#    pip-compile --output-file=requirements.txt requirements.in
+#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
 #
 aiofiles==24.1.0
     # via -r requirements.in
@@ -25,6 +25,8 @@ attrs==25.3.0
     # via aiohttp
 bcrypt==4.3.0
     # via passlib
+black==25.1.0
+    # via -r requirements-dev.in
 blis==1.3.0
     # via thinc
 catalogue==2.0.10
@@ -44,6 +46,7 @@ charset-normalizer==3.4.2
 click==8.2.1
     # via
     #   -r requirements.in
+    #   black
     #   typer
     #   uvicorn
 cloudpathlib==0.21.1
@@ -67,6 +70,8 @@ ecdsa==0.19.1
     # via python-jose
 fastapi==0.116.1
     # via -r requirements.in
+flake8==7.3.0
+    # via -r requirements-dev.in
 frozenlist==1.7.0
     # via
     #   aiohttp
@@ -84,13 +89,17 @@ httpcore==1.0.9
 httptools==0.6.4
     # via uvicorn
 httpx==0.28.1
-    # via -r requirements.in
+    # via
+    #   -r requirements-dev.in
+    #   -r requirements.in
 idna==3.10
     # via
     #   anyio
     #   httpx
     #   requests
     #   yarl
+iniconfig==2.1.0
+    # via pytest
 jinja2==3.1.6
     # via spacy
 joblib==1.5.1
@@ -111,6 +120,8 @@ markupsafe==3.0.2
     # via
     #   jinja2
     #   mako
+mccabe==0.7.0
+    # via flake8
 mdurl==0.1.2
     # via markdown-it-py
 multidict==6.6.3
@@ -122,6 +133,12 @@ murmurhash==1.0.13
     #   preshed
     #   spacy
     #   thinc
+mypy==1.17.1
+    # via -r requirements-dev.in
+mypy-extensions==1.1.0
+    # via
+    #   black
+    #   mypy
 numpy==2.3.2
     # via
     #   -r requirements.in
@@ -133,7 +150,9 @@ numpy==2.3.2
     #   thinc
 packaging==25.0
     # via
+    #   black
     #   pytesseract
+    #   pytest
     #   spacy
     #   thinc
     #   weasel
@@ -141,6 +160,10 @@ pandas==2.3.1
     # via -r requirements.in
 passlib[bcrypt]==1.7.4
     # via -r requirements.in
+pathspec==0.12.1
+    # via
+    #   black
+    #   mypy
 pdf2image==1.17.0
     # via -r requirements.in
 pillow==11.3.0
@@ -148,6 +171,10 @@ pillow==11.3.0
     #   -r requirements.in
     #   pdf2image
     #   pytesseract
+platformdirs==4.3.8
+    # via black
+pluggy==1.6.0
+    # via pytest
 preshed==3.0.10
     # via
     #   spacy
@@ -164,6 +191,8 @@ pyasn1==0.6.1
     # via
     #   python-jose
     #   rsa
+pycodestyle==2.14.0
+    # via flake8
 pycparser==2.22
     # via cffi
 pydantic==2.11.7
@@ -179,8 +208,12 @@ pydantic-core==2.33.2
     # via pydantic
 pydantic-settings==2.10.1
     # via -r requirements.in
+pyflakes==3.4.0
+    # via flake8
 pygments==2.19.2
-    # via rich
+    # via
+    #   pytest
+    #   rich
 pyjwt==2.10.1
     # via -r requirements.in
 pymongo==4.14.0
@@ -189,6 +222,12 @@ pypdf2==3.0.1
     # via -r requirements.in
 pytesseract==0.3.13
     # via -r requirements.in
+pytest==8.4.1
+    # via
+    #   -r requirements-dev.in
+    #   pytest-asyncio
+pytest-asyncio==1.1.0
+    # via -r requirements-dev.in
 python-dateutil==2.9.0.post0
     # via pandas
 python-docx==1.2.0
@@ -208,7 +247,7 @@ pyyaml==6.0.2
     # via
     #   -r requirements.in
     #   uvicorn
-redis==5.0.1
+redis==6.4.0
     # via -r requirements.in
 requests==2.32.4
     # via
@@ -271,6 +310,7 @@ typing-extensions==4.14.1
     #   alembic
     #   anyio
     #   fastapi
+    #   mypy
     #   pydantic
     #   pydantic-core
     #   python-docx
