--- backend/requirements-hybrid.txt
@@ -1,36 +1,0 @@
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-python-multipart==0.0.6
-python-jose[cryptography]==3.3.0
-passlib[bcrypt]==1.7.4
-pymongo==4.6.0
-psycopg2-binary==2.9.9
-redis==5.0.1
-prometheus-client==0.19.0
-
-# Security and validation
-python-magic==0.4.27
-python-magic-bin==0.4.14; sys_platform == 'win32'
-chardet==5.2.0
-cryptography==41.0.7
-
-# Document Processing & OCR
-pytesseract==0.3.10
-pdf2image==1.16.3
-python-docx==1.1.0
-docx2txt==0.8
-PyPDF2==3.0.1
-Pillow==10.1.0
-
-# AI & NLP
-spacy==3.7.2
-nltk==3.8.1
-
-# Cloud LLM Providers
-openai==1.3.0
-anthropic==0.7.0
-google-generativeai==0.3.0
-aiohttp==3.9.0
-
-# Additional utilities
-python-dotenv==1.0.0
