--- backend/requirements-llm-fixed.txt
@@ -1,47 +1,0 @@
-# Local LLM Dependencies for Vanta Ledger
-# Optimized for RTX 3050 GPU and multi-company processing
-
-# Core LLM Framework
-llama-cpp-python==0.2.11
-transformers==4.35.0
-torch==2.2.0
-torchvision==0.17.0
-torchaudio==2.2.0
-
-# Document Understanding
-Pillow==10.0.1
-opencv-python==4.8.1.78
-sentence-transformers==2.2.2
-
-# Hardware Detection and Monitoring
-GPUtil==1.4.0
-psutil==5.9.6
-
-# Database and Caching
-redis==5.0.1
-pymongo==4.5.0
-
-# Utilities
-numpy>=1.26.0
-scipy>=1.11.0
-scikit-learn>=1.3.0
-
-# Image Processing
-opencv-python-headless==4.8.1.78
-
-# Text Processing
-nltk==3.8.1
-spacy==3.7.2
-
-# Performance Optimization
-numba==0.57.1
-joblib==1.3.2
-
-# Monitoring and Logging
-prometheus-client==0.17.1
-structlog==23.1.0
-
-# Development and Testing
-pytest==7.4.2
-pytest-asyncio==0.21.1
-pytest-mock==3.11.1
