--- backend/requirements-minimal.txt
@@ -1,8 +1,0 @@
-# Minimal requirements to resolve import errors
-torch>=2.0.0
-transformers>=4.30.0
-llama-cpp-python>=0.2.0
-redis>=4.0.0
-Pillow>=9.0.0
-psutil>=5.8.0
-GPUtil>=1.4.0
