--- backend/requirements-secure.txt
@@ -1,84 +1,0 @@
-# Secure Requirements for Vanta Ledger
-# Updated to address Dependabot security vulnerabilities
-
-# Core Framework
-fastapi==0.116.1
-uvicorn[standard]==0.35.0
-pydantic==2.11.7
-pydantic-settings==2.10.1
-starlette==0.47.2
-
-# Database
-sqlalchemy==2.0.42
-alembic==1.16.4
-psycopg2-binary==2.9.10
-pymongo==4.14.0
-redis==5.0.1
-
-# Authentication & Security
-python-jose[cryptography]==3.5.0
-passlib[bcrypt]==1.7.4
-python-multipart==0.0.20
-PyJWT==2.10.1
-cryptography==45.0.6
-
-# HTTP & Networking
-requests==2.32.4
-aiohttp==3.12.15
-httpx==0.28.1
-urllib3==2.5.0
-
-# File Processing
-python-multipart==0.0.20
-aiofiles==24.1.0
-Pillow==11.3.0
-
-# Data Processing
-numpy==2.3.2
-pandas==2.3.1
-scikit-learn==1.7.1
-
-# LLM & AI
-torch>=2.0.0
-transformers>=4.30.0
-llama-cpp-python>=0.2.0
-sentence-transformers>=2.2.0
-
-# System & Monitoring
-psutil==7.0.0
-GPUtil>=1.4.0
-
-# Utilities
-python-dotenv==1.1.1
-PyYAML==6.0.2
-click==8.2.1
-rich==14.1.0
-
-# Testing
-pytest==8.4.1
-pytest-asyncio==0.24.0
-httpx==0.28.1
-
-# Development
-black==24.10.0
-flake8==7.0.0
-mypy==1.12.0
-
-# Optional: Document Processing
-# Uncomment if needed
-# pytesseract>=0.3.10
-# pdf2image>=1.16.3
-# python-docx>=0.8.11
-# PyPDF2>=3.0.0
-# spacy>=3.7.0
-
-# Optional: Performance
-# Uncomment if needed
-# bottleneck>=1.3.0
-# numba>=0.58.0
-
-# Optional: Advanced AI
-# Uncomment if needed
-# accelerate>=0.20.0
-# bitsandbytes>=0.41.0
-# optimum>=1.12.0
